Refactor ui folder to delete home folder and move its files.
